Refactoring the verify method
