+++ PolymorphicPseudonymisation/Crypto/EcSdsaSignature.cs
@@ -1,4 +1,6 @@
-﻿using Org.BouncyCastle.Math;
+﻿using System;
+using Org.BouncyCastle.Math;
+using Org.BouncyCastle.Math.EC;
 
 namespace PolymorphicPseudonymisation.Crypto
 {
@@ -6,6 +8,11 @@ namespace PolymorphicPseudonymisation.Crypto
     {
         public EcSdsaSignature(BigInteger r, BigInteger s) : base(r, s)
         {
+        }
+
+        public override void Verify(ECPoint publicKey, ECPoint g, byte[] message)
+        {
+            throw new NotImplementedException();
         }
     }
 }
