+++ PolymorphicPseudonymisation/Crypto/Signature.cs
@@ -8,13 +8,13 @@ namespace PolymorphicPseudonymisation.Crypto
 {
     public abstract class Signature
     {
-        private readonly BigInteger r;
-        private readonly BigInteger s;
+        protected BigInteger R { get; set; }
+        protected BigInteger S { get; set; }
 
         protected Signature(BigInteger r, BigInteger s)
         {
-            this.r = r;
-            this.s = s;
+            R = r;
+            S = s;
         }
 
         public static Signature Create(string objectIdentifier, BigInteger r, BigInteger s)
@@ -27,40 +27,39 @@ namespace PolymorphicPseudonymisation.Crypto
             };
         }
 
-        public void Verify(ECPoint publicKey, ECPoint g, byte[] message)
-        {
-            Guard.AssertNotNull(publicKey, nameof(publicKey));
-            Guard.AssertNotNull(g, nameof(g));
-            Guard.AssertNotNull(message, nameof(message));
-
-            if (r.BitCount > 320 || r.CompareTo(BigInteger.Zero) <= 0 || s.CompareTo(BrainpoolP320R1.Q) >= 0)
-            {
-                throw new CryptoException("Invalid signature");
-            }
-
-            var q = g.Multiply(s).Add(publicKey.Multiply(r)).Normalize();
-            if (q.IsInfinity)
-            {
-                throw new CryptoException("Invalid signature");
-            }
-
-            var md = Sha384.Instance;
-            var encodedXCoordinate = q.AffineXCoord.GetEncoded();
-            md.TransformBlock(message, 0, message.Length, message, 0);
-            md.TransformFinalBlock(encodedXCoordinate, 0, encodedXCoordinate.Length);
-
-            // Use only 320 MSB
-            var hash = Arrays.CopyOfRange(md.Hash, 0, 40);
-            var v = new BigInteger(1, hash);
-            if (!r.Equals(v))
-            {
-                throw new CryptoException("Invalid signature");
-            }
-        }
+        public abstract void Verify(ECPoint publicKey, ECPoint g, byte[] message);
 
         public override string ToString()
         {
-            return $"{r.ToString()}, {s.ToString()}";
+            return $"{R.ToString()}, {S.ToString()}";
         }
+
+        protected void AssertInputRequirements()
+        {
+            if (R.BitCount > 320 || R.CompareTo(BigInteger.Zero) <= 0 || S.CompareTo(BrainpoolP320R1.Q) >= 0)
+            {
+                throw new CryptoException("Invalid signature");
+            }
+        }
+
+        protected static BigInteger Get320MostSignificantBytes(System.Security.Cryptography.SHA384 md)
+        {
+            Guard.AssertNotNull(md, nameof(md));
+
+            var hash = Arrays.CopyOfRange(md.Hash, 0, 40);
+            var v = new BigInteger(1, hash);
+
+            return v;
+        }
+
+        protected void AssertsSignaturesAreEqual(BigInteger v)
+        {
+            if (!R.Equals(v))
+            {
+                throw new CryptoException("Invalid signature");
+            }
+        }
+
+
     }
 }
